src/strand_agent_poc/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
certifi==2025.8.3
click==8.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
mcp==1.13.0
pip==25.2
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
python-dotenv==1.1.1
python-multipart==0.0.20
referencing==0.36.2
rpds-py==0.27.0
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.47.2
typing-extensions==4.14.1
typing-inspection==0.4.1
uvicorn==0.35.0
strands-agents>=1.0.0
strands-agents-tools>=0.2.0
uv
boto3
bedrock-agentcore
bedrock-agentcore-starter-toolkit
